fix: enable global theme switching using context API
