+++ src/contexts/AppContext.jsx
@@ -1,17 +1,34 @@
-import { createContext, useState, useEffect } from "react";
+// src/context/AppContext.jsx
+import React, { createContext, useEffect, useState } from "react";
 
 export const Context = createContext();
 
 export const AppProvider = ({ children }) => {
   const [token, setToken] = useState(null);
 
+  // ✅ Theme state
+  const getInitialTheme = () => localStorage.getItem("theme") || "light-theme";
+  const [theme, setTheme] = useState(getInitialTheme());
+
+  const toggleTheme = () => {
+    const newTheme = theme === "dark-theme" ? "light-theme" : "dark-theme";
+    setTheme(newTheme);
+    localStorage.setItem("theme", newTheme);
+  };
+
+  useEffect(() => {
+    document.body.className = theme;
+  }, [theme]);
+
   useEffect(() => {
     const storedToken = localStorage.getItem("token");
-    if (storedToken) setToken(storedToken);
+    if (storedToken) {
+      setToken(storedToken);
+    }
   }, []);
 
   return (
-    <Context.Provider value={{ token, setToken }}>
+    <Context.Provider value={{ token, setToken, theme, toggleTheme }}>
       {children}
     </Context.Provider>
   );
